Move switching course instances seed
